saga-28/app/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = []


[tool.poetry.dependencies]
django = "~3.2"
django-health-check = "^3.14.2"
djangorestframework = "~3.12"
psycopg2-binary = "^2.8.6"
python = ">=3.8,<3.9"
djangorestframework-simplejwt = "^5.0.0"
requests = "^2.27.1"
django-solo = "^2.0.0"
